pyproject.toml
[tool.poetry]
name = "terrable"
version = "0.2.0"
description = "Manage private shared terraform modules deployed to S3."
authors = ["Scott Ernst <swernst@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://gitlab.com/rocket-boosters/terrable"
repository = "https://gitlab.com/rocket-boosters/terrable"
documentation = "https://gitlab.com/rocket-boosters/terrable"
keywords = ["terraform", "s3", "modules"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Utilities",
  "Typing :: Typed"
]
exclude = ["lobotomy/tests"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.16.16"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-runner = "*"
pytest-cov = "*"
flake8 = "*"
mypy = "*"
coverage = "*"
black = { version = "*", allow-prereleases = true }
flake8-black = "*"
taskipy = "*"
lobotomy = "^0.3.7"
aok = "^0.3.0"
pydocstyle = "^6.1.1"
radon = "^5.0.1"
yamllint = "^1.26.1"

[tool.poetry.scripts]
terrable = 'terrable:main'

[tool.taskipy.tasks]
# https://github.com/illBeRoy/taskipy
black = "black ."
black_lint = "black . --check"
mypy = "mypy . --install-types --non-interactive --ignore-missing-imports"
flake8 = "flake8 ."
pydocstyle = "pydocstyle ."
yamllint = "yamllint ."
radon = "radon cc . && radon mi ."
test = "pytest . --cov-report=term-missing --cov=."
format = "task black"
lint = "task black_lint && task flake8 && task mypy && task pydocstyle && task radon && task yamllint"
check = "task format && task lint && task test"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
